backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyMuPDF
python-multipart
plotly
numpy
kaleido